refactor api access
