+++ src/hooks/useCategories.js
@@ -1,13 +1,5 @@
-import axios from "axios";
 import { useEffect, useState } from "react";
-
-const fetchCategories = () => {
-  return axios
-    .get("https://nc-games-portfolio.herokuapp.com/api/categories")
-    .then(({ data: categories }) => {
-      return categories;
-    });
-};
+import { fetchCategories } from "../api";
 
 export const useCategories = () => {
   const [categories, setCategories] = useState("all");
