+++ src/hooks/useReviews.js
@@ -1,14 +1,5 @@
-import axios from "axios";
 import { useEffect, useState } from "react";
-
-export const fetchReviews = (category) => {
-  let queryString = "https://nc-games-portfolio.herokuapp.com/api/reviews";
-  if (category) queryString += `?category=${category}`;
-
-  return axios.get(queryString).then(({ data: reviews }) => {
-    return reviews;
-  });
-};
+import { fetchReviews } from "../api";
 
 export const useReviews = (category) => {
   const [reviews, setReviews] = useState([]);
