+++ src/hooks/useScoreIncrease.js
@@ -1,12 +1,5 @@
 import { useState } from "react";
-import axios from "axios";
-
-const increaseScore = (review_id) => {
-  return axios.patch(
-    `https://nc-games-portfolio.herokuapp.com/api/reviews/${review_id}`,
-    { inc_votes: 1 }
-  );
-};
+import { increaseScore } from "../api";
 
 export const useScore = (review_id) => {
   const [clickedScore, setClickedScore] = useState(0);
